Update vscode debugging config in docs (#40274)
### Description of work
In all honesty, this change is not required, since the previous vscode
configuration was already working.
I've only simplified a little bit.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
Copy and paste the configuration into your `launch.json` file in VS Code
and check debugging in C++ is working.