Pre-populated fields in GenericDialog (#40132)
### Description of work
The primary objective of this work is to enable convenient
pre-population of properties in user-interface dialogs when used with
Python algorithms. Such pre-populated dialogs can then be utilized using
only the `GenericDialog` interface, and do not require the
implementation of a custom dialog for this reason alone.

### Explanation of work
Dependent properties in a dialog are set based on the values of upstream
properties. For example, if an `InputWorkspace` property is specified,
several other properties may take their initial values from the
input-workspace logs. There are several different ways to facilitate
this type of behavior, and for the present purposes these were
distinguished largely by how they treat the case where the end user has
already entered a value for the dependent property.

There are two general cases:

1. Change values of dependent properties regardless of whether or not
the user has already entered a value for these properties. This case is
implemented by the _existing_ `IPropertySettings`-derived class
`SetValueWhenProperty`. The first step in the work of this PR was simply
to export this `IPropertySettings` variant to the Python API.

2. Change values of dependent properties only when those properties have
not already been modified by the end user. This case is implemented by a
new `IPropertySettings`-derived class `SetDefaultWhenProperty`, which is
also exported to the Python API. For this particular case, note that the
_actual_ default values of the properties are not modified, and can be
recovered by resetting the `IPropertyManager` that the dialog is
serving.

Several adjustments have been made to the `IPropertySettings` interface:

* A new `dependsOn` virtual method is provided. This method returns a
list of the names of any upstream properties that the current property
depends on. This method raises an exception when circular dependencies
are detected. (This method is intended for use by specialized versions
of `IPropertySettings`, or by advanced dialog features. This method was
required by the initial implementation of `SetDefaultWhenProperty`, but
is no longer required by the current treatment. This change was left in
as a feature improvement.)
  
* The external-facing methods (in C++) are modified to prefer the use of
property names, rather than the property pointers themselves. This
change is made to support the possible case where `applyChanges`
actually replaces a property on the `IPropertyManager`. (This form was
required by an initial implementation of `SetDefaultWhenProperty`, but
is no longer required by the current treatment. This change was also
left in as a feature improvement.)

* The number of methods exported by the interface has been extended,
with a few methods added to facilitate unit testing from Python. By
exporting these methods for testing we are able to more completely test
more complicated features of the export interface itself -- this is
particularly relevant to the export of generalised `std::function`
types, which is now fully supported.

Templates are now provided supporting the export to Python of
generalised `std::function`-based types.

**Apart from the release note, an additional follow-on story will add
more comprehensive documentation covering how to use these changes in
development work.**

Work item link: [EWM
#13621](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=13621)

### To test:

Unit tests have been added to cover the new `IPropertySettings`
variants. In order to test the features of either of the
`IPropertySettings` variants, a script similar to that included below
may be used from Mantid workbench. Please note that the `criterion`
function or bound-method (it only needs to be a `Callable`) passed to
`SetValueWhenProperty` is
`Callable[[str, str], str]`, but for `SetDefaltWhenProperty` is 
`Callable[[IPropertySettings, Property, Property], bool]`.  
For `SetValueWhenProperty`, the return value is the new value for the
current property, which might be the same as the old value. For
`SetDefaultWhenProperty` the return value indicates whether or not the
current property has been modified.

**PLEASE see the attached test script!**

[prePopulateScript.py](https://github.com/user-attachments/files/23168850/prePopulateScript.py)


<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->

<!--  GATEKEEPER: When squashing, remove the section from HERE...  -->
---

### Reviewer

**Your comments will be used as part of the gatekeeper process.**
Comment clearly on what you have checked and tested during your review.
Provide an audit trail for any changes requested.

As per the [review
guidelines](http://developer.mantidproject.org/ReviewingAPullRequest.html):

- Is the code of an acceptable quality? ([Code
standards](http://developer.mantidproject.org/Standards/)/[GUI
standards](http://developer.mantidproject.org/Standards/GUIStandards.html))
- Has a thorough functional test been performed? Do the changes handle
unexpected input/situations?
- Are appropriately scoped unit and/or system tests provided?
- Do the release notes conform to the
[guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)
and describe the changes appropriately?
- Has the relevant (user and developer) documentation been
added/updated?

### Gatekeeper

As per the [gatekeeping
guidelines](https://developer.mantidproject.org/Gatekeeping.html):

- Has a thorough first line review been conducted, including functional
testing?
- At a high-level, is the code quality sufficient?
- Are the base, milestone and labels correct?
<!--  GATEKEEPER: ...To HERE  -->

---------

Co-authored-by: Kort Travis <traviska@ornl.gov>