Refactor Reflectometry tests (#40230)
### Description of work
Overdue refactor of the reflectometryauto3 test suite. 
There were many redundant lines, this was a bit painful to do, but I
think it's more clear now and easy to navigate.

There was also a wrong extension for the reference to a unit test file
(POLREF0001466.RAW) that was making the test to take much longer than
they should, I have fixed this for the affected tests.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes [#38837](https://github.com/mantidproject/mantid/issues/38837).
<!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
- Code check.