Application of multiple `IPropertySettings`
Multiple ``IPropertySettings`` can now be attached to a single property, by applying either ``IPropertyManager::setPropertySettings`` or the ``Property::setSettings`` method of the property itself, multiple times.  ``Property::getSettings`` now returns a *vector* of settings, which will be *empty* in the default no-settings case.  It is expected that up to one instance of each of the ``EnabledWhenProperty`` and ``VisibleWhenProperty`` settings types would be attached to a given property, in combination with any number of instances of other ``IPropertySettings``-derived types, although this behavior is not strictly enforced.

This PR includes the following changes:

  * `Property::getSettings`, `setSettings`, `clearSettings`, `IPropertyManager::setPropertySettings` -- return, append to, or clear a `std::vector<std::unique_ptr<IPropertySettings const>>`;

  * With respect to the property's enabled and visibility states (on its owner widget), any setting in the vector of settings returning ``isEnabled`` or ``isVisible`` values of ``false``, results in disabling or hiding the property's widget. This behavior is centralized and enforced using the new `IPropertyManager::isPropertyEnabled`, `IPropertyManager::isPropertyVisible` methods.  The fact that these could not usefully be `IPropertySettings` methods is related to issues with the object-oriented design of the current `IPropertySettings` system, and the necessity to support legacy behavior in the codebase;

  * Where appropriate, `IPropertySettings` methods are now `const`;

  * New unit tests have been added for `AlgorithmPropertiesWidget`.  At present these tests have incomplete coverage.  Current tests focus on required `hideAndDisableProperties` and `IPropertySettings`-related behaviors.