Fix test values for height and sigma in FindEPP
Centre unchanged and chi-squared of fit same as on main to within 6 decimal points (note the unit test asserts chi-squared to 4 decimal points - and that line did not cause a failure).
https://github.com/mantidproject/mantid/blob/c91f05d6ee4a4c60fc552671bae075047ff1cb7e/Framework/Algorithms/test/FindEPPTest.h#L145

New parameters are negligibly closer to actual peak parameter used to simulate data, but the fit is unrealistic anyway as it doesn't account for the FlatBackground present in the data but not in model.