Update corelli calibration test_bank
Improvements to fitting Gaussian peaks seem to have produced fewer 'outlier' tubes that require masking.
The cost function of the fits is better in all cases. In the tubes that were previously masked but now not identified as outliers, the deviation in the vertical positions from the average for all tubes in the bank (see function criterion_peak_vertical_position) was consistent with other valid tubes.