New algorithm for computing spectral moments (#40323)
This algorithm is a step in calculating sum rules for inelastic
scattering. It takes MDEvent workspaces, with one of the dimensions
being DeltaE, and multiplies each event by DeltaE^n. To calculate sum
rules, one needs to BinMD/MDNorm the data in energy, then sum along that
axis.


**Report to:** Martin Mourigal, Kemp Plumb


### To test:

The documentation test script is a good start

---------

Co-authored-by: Kyle Qianli Ma <41385391+KyleQianliMa@users.noreply.github.com>