Create a SystemTest Runner for Use with `ctest` (#40296)
- Adds the full suite of system tests to CMake so that `ctest` can run
them.
- Adds labels for unit and system tests such that the one that should be
run can be selected `-L UnitTest/SystemTest` when running `ctest`

---
- **NOTE: Does NOT replace the call to `systemtest`/`systemtest.bat` in
the CI run config. For now, just the implementation of the runner to be
checked. Intended to reduce the size of the PR, which can be more easily
tested. Not all of the tests will currently pass using `ctest`.**

Refs #34850

### To test:
1. Re-run CMake.
2. Run a subset of unit tests using: `ctest -R <whatever> -L UnitTest`
3. Run a subset of system tests using `ctest -R <whatever> -L
SystemTest`
4. Run a subset of all tests using `ctest -R <whatever>`
5. Ensure that `systemtest -R <whatever>` still works. 


Co-authored-by: Donald Chung  <DonaldChung-HK@users.noreply.github.com>

---------

Co-authored-by: Jonathan Haigh <jhaigh0@users.noreply.github.com>