Side by side projection in new Instrument View (#39949)
Adds the "side-by-side projection" to the new Instrument View.

<img width="1459" height="787" alt="image"
src="https://github.com/user-attachments/assets/63443735-f95f-4fbd-9aab-417addc0f7bd"
/>

Closes #38429.

I've moved the code I could out from the old class, `PanelsSurface` into
a new calculator class, `PanelsSurfaceCalculator`, which allowed me to
reuse it in the new Instrument View.

The legend is now draggable and resizeable.

Detectors in grids will be drawn as such, and any other types of
detector will be grouped into a separate grid, with their position in
the grid abstract and not related to their physical position (it will
arrange them as close as possible to a square).

The panel rotation problem mentioned by Richard in #38429
[here](https://github.com/mantidproject/mantid/issues/38429#issuecomment-3069923041)
seems to be fixed.

The ability to flip and rotate panels is something we can add in a
future PR, see
[here](https://github.com/mantidproject/mantid/issues/38429#issuecomment-3069961168)

### To test:

The side-by-side option should appear as one of the projection options.

The old instrument view (especially side-by-side) also needs checking
since I refactored it a bit.