Index column in `CreateDetectorTable` is now int (#40336)
`Index` was a double, now it is an `int`. The workspace index is
originally stored as an `int_32` so no need to cast it to a double.

I spotted this because the `columnArray` method on a table workspace was
returning a `numpy` array of type `float` for `Index`, which resulted in
workspace indices in the new Instrument View appearing as floating point
numbers.

### To test:

```python
ws = LoadEmptyInstrument(InstrumentName='SXD')
d = CreateDetectorTable(ws)
ws_index = d.columnArray('Index')
print(type(ws_index[0]))
```
The `columnArray` method is not in 6.14 but it is in the nightly.