Add ability to request TimeAtSampleElastic information (#40341)
To prepare for fast event filtering, the constants for calculating the
time-of-flight of the neutron when it was at the sample position need to
be calculated and made available. This continues from #40243 which put
the calcutions into API so DataHandling could access them.

Refs #40189 and
[EWM13754](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20(Change%20Management)#action=com.ibm.team.workitem.viewWorkItem&id=13754)

### To test:

Unfortunately, this change adds new functionality to an inner class to
prepare it for use in a future PR. The additional unit test might be all
that one can do.

*This does not require release notes* because the follow-on PR to add
fast event filtering functionality to `AlignAndFocusPowderSlim` will add
the notes.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>