Add option for flipper analyzer calculation in FlipperEfficiency and SANSISISPolarizationCorrections algorithms (#40253)
### Description of work
Ditto. 
The equation for the flipper analyzer is very similar to the flipper
polarizer, just changing the cross terms (01->10, 10->01), I have
refactor a bit the `FlipperEfficiency` algorithm to accomodate this and
the choice for which flipper to use in the algorithm is accessed through
a `StringListValidator` with the choices `[Polarizer, Analyzer]`.
For completion, it is possible to calculate the efficiency of the second
flipper in the SANSISISPolarizationCorrections algorithm if the
`SecondFlipperEfficiency` checkbox is ticked on.
I have absorbed #40208 into this one, as it is a very minimal change. 

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes #37293 
Closes #40208

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
- Build docs, and check new algorithm description is clear and release
notes are well formatted.
- Build this branch
- Try test example in the docs, changing to `Analyzer`
- To test #40208, open `SANSISISPolarizationCorrections` in the
algorithm dialog from mantid workbench. When correction mode
`Calibration` is selected, the property `ScatteringRuns` should be
greyed out, when `Correction` mode is selected, the properties
`Transmissionruns`, `DirectBeamRun` and `DepolarizedCellRun` should be
greyed out. NOTHING should be greyed out when selecting
`CalibrationAndCorrection` mode.
<!-- Include sufficient instructions for someone unfamiliar with the
application to test.
Ok to refer back to instructions in the issue.
-->

<!-- REMEMBER:
- Add labels, milestones, etc.
- Ensure the base of this PR is correct (e.g. release-next or main)
- Add release notes in separate file as per
([guidelines](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)),
or justify their absence:
*This does not require release notes* because <fill in an explanation of
why>
-->