Tell agents what version of c++ and python are supported (#40388)
To aid AI development, inform the agents what version of C++ and Python
we are using so they can make more informed decisions.

There is also a small change to the rules for running the github action
for setting up an environment for copilot. This removes duplicate runs
in PRs that modify the job.

There is no associated issue and this doesn't need release notes because
it is a change to the instructions for AI agents.

### To test:

Review the change in text and see that the `copilot-setup-steps` job is
only run once.