Expose number of solid angle cylinder slices to SANS state (#40242)
### Description of work

This PR adds the option to the SANS user file to specify a number of
cylinder slices to be used when calculating the Solid Angle.

This value is set to 11 by default, as opposed to the 10 used as default
in the underlying algorithms. Because of this a large number of
reference files for system tests have had to be updated.

Plots comparing the results can be found here:
`\\Olympic\babylon5\Public\miallewis`

To make the changing of a large number of reference files easier, I
added the ability to specify an environment variable
`GENERATE_REFERENCE_FILES_ON_FAILURE=1`, that essentially reroutes the
mismatch file into the reference data folder, ensuring it is named the
same as the original reference file.

Closes #40221 <!-- One line per closed issue. -->