Measure angle between two picked detectors (#40300)
Measure the angle between the Q lab directions of two picked detectors.
The old interface required a peak in order to measure this, but since
all we are interested in is the angle between the Q lab vectors, we
don't need peaks. The peak will tell you the magnitude of the Q lab
vector, but we only need the direction.

The old interface reported the angle in degrees so I did the same here.

Fixes #40044 

### To test:

Compare the angle displayed in the old Instrument View with the new one.
In the new IV, you just select any two detectors and it will tell you
the angle. In the old one you have to click the 'compare' button on the
peaks tab, and your detectors need to have peaks.