Improve user file validation in SANS gui (#40353)
### Description of work
It was reported in #38118 that setting a directory on the user file path
can led to the interface being unable to load.
Although it is not clear how can this be done within the interface ( I
haven't been able to do this myself), it is possible to set a directory
externally by meddling on the sans specific qsettings.
This PR checks for that condition when opening the interface form the
presenter and launches an error box describing the problem, additionally
the user file path is reset and the interface opens normally.
<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes #38118. <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:
- Run the script on #38118 to set a directory as a user file. 
- Try to launch the sans gui. It should open. 

Co-authored-by: Caila Finn <47181718+cailafinn@users.noreply.github.com>