Only run jobs in mantidproject/mantid repository (#40416)
Since forks likely won't need some of the maintenance running, modify
the condition for running the jobs to only be in `mantidproject` owned
repositories.

There is no associated issue.

### To test:

A code review is the appropriate way to see, but copying this PR into a
fork repository would be a more complete test.