Suppress rendering during plotter update
Introduced to fix issues with wrong transforms being calculated and
unpleasant flahes that were happening because the plotter was being
updated for every mesh added to it.

Instead of many incremental changes to the plotter it's better to
disable rendering, do all of the changes and then make those changes
render those changes to the plotter, so they appear all in one go.

Co-authored-by: James Clarke <james.clarke@stfc.ac.uk>