ReorientUnitCell algorithm (#40327)
This PR introduces the ReorientUnitCell algorithm, which reorients a
unit cell to be most aligned with goniometer axes by finding the
symmetry transformation that maximizes the trace of the UB matrix. The
algorithm evaluates all symmetry operations of the crystal's point group
to find the orientation closest to the identity matrix.

Implements [EWM
13648](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=13648)
Companion to #40328

### To test:
- Build the documentation with `ninja docs-html` to make sure the
description of algorithm `ReorientUnitCell` makes sense.

---------

Signed-off-by: Jose Borreguero <borreguero@gmail.com>