Precommit job to update the pixi toml to have the same dependencies as the mantid-developer package (#40403)
### Description of work

This PR
1. Adds a copy of the dependencies of `mantid-developer` to the base
level dependencies of the pixi manifest file
- This is being done as part of #40294 so that we can use a local
version of mantid-developer to run in the ci for some tasks. A later PR
will do this conversion and add a pixi environment for using
`rattler-build` (and related tools) for other ci jobs.
2. Adds a local precommit script which will update the pixi toml (using
`pixi add / remove`) to reflect changes made to the `mantid-developer`
conda recipe
- This is to combat the problem which the `mantid-developer` package was
intended to prevent, which is dependency pins being stored in multiple
locations
- The script will not work in the other direction (so if you edit
pixi.toml, you're on your own) but it will check to see if the
environments match, failing if they don't. This should keep us safe from
mismatched pins.

Q: Do we need `mantid-developer` any more anyway?
A: We plan to keep publishing it as it is very convenient for developers
using conda / mamba to create their development environments. The
alternative would mean everyone have to switch to using pixi which is a
relatively new tool and requires a bit of a different way of working.

### To test:

You can run pre-commit with `.git/hooks/pre-commit` (you will need to
`git add` the changes you are testing first though).
Some suggested tests:

- Change a pin in the `conda_build_config.yaml` file (must be one used
by `mantid-developer`). You will need to make sure you change the pin to
something that makes sense (package that exists, won't cause conflicts).
 - Add / remove a dependency from `mantid-developer/recipe.yaml`
- Move a dependency from `mantid-developer/recipe.yaml` to / from a
specific os
- Try making changes to the `pixi.toml` pins. The pre-commit should fail
and tell you where there is a mismatch.

Happy for suggestions to be made on code structure etc. there are
probably a lot of ways to solve this problem.