Corrections to automerge CI skipping (#40434)
In #40416 there was an attempt to skip automerging `ornl-next` and
`release-next` into `main`. The selection was done in a way that they
don't appear to run (see example on
deeb84997194248e474796b3635e93407a1d38ef). This copies the selection
mechanism used in `update-pixi-lockfile.yml` to bring it back.

There is no associated issue.

### To test:

The only real way to test this is to merge it and get it into
`ornl-next`.

*This does not require release notes* because it is fixing a bug in
automated code maintenance.