Add Basic masking to New instrument view (#40397)
This PR adds the bare-bones for masking in the new instrument view.
Currently only one shape is supported (a circle) and I included a list
widget that allows the user to pick between masks, combine them and
finally export them (to workspace, file or burn the mask permanently
into the workspace).

### To test:
Use the new widget to add a circle, apply several masks, choose some of
the masks and finally export the masks into a workspace, save the masks
to a file and burn the masks permanently into the workspace.

---------

Co-authored-by: James Clarke <james.clarke@stfc.ac.uk>