Fix pixi pre-commit bug and fix the boost pin (#40458)
### Description of work

A bug existed in the pixi pre-commit script added recently such that it
couldn't cope with a package (such as `libboost-devel`) which is
included on all platforms, but has a different version specified for
each.

Also pulled in the commits to fix rstcheck from this pr
https://github.com/mantidproject/mantid/pull/40456

### To test:

Code review.
Could do some test changes to the conda recipe to check that they're
read properly.


Co-authored-by: Andrei Savici <saviciat@ornl.gov>