Avoid duplicate modal window on sans gui when changing line edits (#40411)
### Description of work
This was found whilst working on : #39596 
When entering a wrong setting on the settings tab of the sans gui, for
example, setting 'a' on the Radius Limit line edit, which expects
something that can be cast into float, sends a warning message to the
user as a modal qmessagebox.
If the wrong setting is typed on the line edit, and the line edit goes
out of focus (clicking somewhere else or pressing tab), a single modal
window appears. But if the return key is pressed on the line edit, the
signal is sent twice, once when the return key is pressed, and a second
one when the modal window appears on top and takes the focus away from
the line edit.
This is not much of an issue on linux and windows (you can close the two
modal windows and continue), but it on macosx the modal windows appear
to work a bit differently, and it can cause the interface to be blocked
when closing the warning boxes.
I have added a simple fix in this PR. There may be a better way to fix
this, I'm open for suggestions on this PR, but for the time being it
serves to avoid a possible crash on macos.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

No associated issue. <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

- Try setting wrong inputs on the mask line edits of the settings tab of
the sansgui interface, expect to have only one warning box.