Add ability to turn off some of the local repository mirrors (#39957)
Rather than relying on network and such to fail quickly from servers
that cannot be seen, add options to only include selected local mirrors.
Since this happens at the cmake configuration stage, only the desired
mirrors will be added to the list to search. The new mechanism defaults
to using all mirrors.

Nothing was changed in buildscripts because it requires a mechanism to
determine which build server is being used.

There is no associated issue, but it is associated with
[EWM13274](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20(Change%20Management)#action=com.ibm.team.workitem.viewWorkItem&id=13274)

### To test:

Try the various options `all`, `none`, `ornl`, and `ral` and look for
the cmake message about which mirrors are used. One can also search the
`CMakeCache.txt` for the value of `ExternalData_URL`.