Peak overlays now fully working with multi-detector IDs per spectrum (#40415)
This fixes a bug with the peak overlay in the new Instrument View, where
the overlay would not display the peaks if the detector ID in the peaks
workspace is not the one used by the Instrument View. This can happen if
an instrument has multiple detectors in a single spectra, e.g. WISH.

### To test:

Pick some peaks on WISH using the old Instrument View, and view them in
the new Instrument View.