Instrument View - Apply camera cache from presenter (#40462)
### Description of work
Changed logic of caching camera positions since caching needs to happen
after transform is applied.
Fixes current bug where cache happens before transforms and resets to an
incorrect position.
Also fixed bounding box to adjust to whatever the current mesh is,
before the bounding box was fixed and transforms could push detectors
outside of it, which was really not good.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

- Open instrument view and check that all detectors are correctly
positioned on the screen
- Do some zooming in and reset the camera
- Add a circle mask widget, zoom in, and reset the camera before
applying the mask
- Tick/untick aspect ratio and repeat all of the steps above