Rebase `FileID` to inherit from `UniqueID` (#40474)
The Nexus refactor work left `Nexus::File` using the HDF5 C-API, which
handles objects through identifying integers. Close operations must be
called on these IDs, or else memory leaks can occur.

At the end of the Nexus refactor, in PR #39732 created a special class
`FileID` to house the ID corresponding to the file, to ensure file
closure on exit.

After a memory leak was discovered in `Nexus::File::getSlab()`, a
similar class, `UniqueID` was created in PR #40368, to handle more
general HDF5 handles. Then in PR #40464, this class was made more
official and used in more situations.

Since the APIs of `UniqueID` and `FileID` are similar, and they serve
similar uses, it is reasonable to think `FileID` can be made a special
type of `UniqueID`. This PR simplifies things by making `FileID` the
specialization of `UniqueID` with a `H5Fclose` deleter.

### To test:

This is a refactor. Ensure all tests pass. Note there are existing tests
of `FileID` inside `NexusFileTest`.

Please run the system test VanadiumAndFocusWithSolidAngleTest,
```
ninja Framework && ./systemtest -R VanadiumAndFocusWithSolidAngleTest
```

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>