Add more complete warning log when Mslice is not available on Mantid (#40470)
### Description of work
Part of mantidproject/mslice#1134

As MSlice is an optional dependency of Mantid for conda installations
since 6.14, it is possible that a user with a conda installation may
want to launch mslice from mantid. As of now, the warning message is
simply `Mslice is not available`, I have added a small description and a
link to the mslice installation instructions to provide more context to
the user that is trying to launch `MSlice` from a conda install of
Mantid and is not aware of this.

<!-- Please provide an outline and reasoning for the work.
If there is no linked issue provide context.
-->

Closes mantidproject/mslice#1134. <!-- One line per closed issue. -->

<!-- If issue raised by user. Do not leak email addresses.
**Report to:** [user name]
-->

### To test:

- Build this branch and try to launch mslice from mantid (you shouldn't
have mslice installed in your conda environment), a more verbose warning
message should appear in the Mantid messages logger.