Make `SharedID` for `Nexus::File` (#40485)
Follows on from the Nexus refactor. Due to the use of the HDF5 C-API
within `Nexus::File`, there were several occasions that could lead to
memory leaks. Previously, this was addressed by the creation of
`UniqueID`, mirroring `unique_ptr`, which ensures only one owner of HDF5
handles and that the correct close method is called on scope exit. This
was handled in PRs:
- https://github.com/mantidproject/mantid/pull/40368
- https://github.com/mantidproject/mantid/pull/40464
- https://github.com/mantidproject/mantid/pull/40474

To prevent memory leaks or premature file close errors across multiple
owners opening the same file, `Nexus::File` was given a `shared_ptr`
that pointed to a `UniqueID<&H5Fclose>` (that is, a file handle).
However, this was kind of a kludge to allow multiple ownership of a
class designed for single ownership.

This PR creates a `SharedID`, similar to `UniqueID`, but allowing
multiple ownership. Owners are tracked through a counter, and all
co-owners have a pointer to this counter. When the number of leashes
drops to zero, the corresponding closer method is called on the managed
ID. Tests help ensure this is thread safe.

### To test:

Build and run the `VanadiumAndFocusWithSolidAngleTest`:
```
ninja Framework && ./systemtest -R VanadiumAndFocusWithSolidAngleTest
```

Build and run the Multithreaded test:
```
ninja Framework && ./systemtest -R MultiThreadedLoadNeXusTest
```

Check the new tests, especially the tests of thread safety, to ensure
the logic makes sense.

Try building and running this script:
``` python
# import mantid algorithms, numpy and matplotlib
from mantid.simpleapi import LoadNexusProcessed

dir = # path to biosans test data needs to go here
file1 = f'{dir}/CG3_127_5532_mBSub.h5'
file2 = f'{dir}/CG3_127_5562_mBSub.h5'
for i in range(3):
    ws1 = LoadNexusProcessed(Filename=file1, OutputWorkspace='sample', LoadHistory=False)
    ws2 = LoadNexusProcessed(Filename=file2, OutputWorkspace='Main_buffer', LoadHistory=False)
    ws1.delete()
    ws2.delete()
```
to check for no memory leaks during the load process. Total memory usage
should stay roughly constant, even if the `for`-loop is re-run multiple
times.