Application of multiple `IPropertySettings` (#40487)
Multiple ``IPropertySettings`` can now be attached to a single property,
by applying either ``IPropertyManager::setPropertySettings`` or the
``Property::setSettings``, multiple times. ``Property::getSettings`` now
returns a *vector* of settings, which will be *empty* in the default,
no-settings case.

It is expected that up to one instance of each of the
``EnabledWhenProperty`` and ``VisibleWhenProperty`` settings types would
be attached to a given property, in combination with any number of
instances of other ``IPropertySettings``-derived types, although this
behavior is not strictly enforced.

### Explanation of work

This PR includes the following changes:

* `Property::getSettings`, `setSettings`, `clearSettings`,
`IPropertyManager::setPropertySettings` -- return, append to, or clear a
`std::vector<std::unique_ptr<IPropertySettings const>>`;

* With respect to the property's enabled and visibility states (on its
owner widget), any setting in the vector of settings returning
``isEnabled`` or ``isVisible`` values of ``false``, results in disabling
or hiding the property's widget. This behavior is centralized and
enforced using the new `IPropertyManager::isPropertyEnabled`,
`IPropertyManager::isPropertyVisible` methods. The fact that these could
not usefully be `IPropertySettings` methods is related to issues with
the object-oriented design of the current `IPropertySettings` system,
and the necessity to support legacy behavior in the codebase;

  * Where appropriate, `IPropertySettings` methods are now `const`;

* New unit tests have been added for `AlgorithmPropertiesWidget`. At
present these tests have incomplete coverage. Current tests focus on the
required `hideAndDisableProperties` and `IPropertySettings`-related
behaviors of this current PR.

Link: [EWM
#14018](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=14018)

### To test:
New unit tests have been added to cover the changes. (See especially
`qt/widgets/common/test/AlgorithmPropertiesWidgetTest.h`.) A simple test
script
[prePopulateScript_multi.py](https://github.com/user-attachments/files/24163076/prePopulateScript_multi.py)
is attached. This script should serve as a starting point to demonstrate
the capabilities of the new changes. (The current script changes a
property's *dynamic default* value, using thresholds based on the values
of several upstream properties. Additional tests to try might involve
adding an `EnabledWhenProperty`, or doing something like taking a sum of
the upstream values. :) )

---------

Co-authored-by: Kort Travis <traviska@ornl.gov>