More tests of `UniqueID` and `SharedID` (#40509)
In the previous PR #40485, a `SharedID` object was created, to
complement the `UniqueID` created before.

This provides additional unit tests of `UniqueID` and `SharedID`.

### To test:

This PR is adding more tests, so make sure they all pass.

This does not require release notes because it is an internal change
only visible to the test suite.