Use ``dorny/paths-filter`` in Doxygen workflow (#40501)
Ensures that the workflow always runs, but will skip the doxygen build
if no changes are made to relevant files. This will allow us to mark the
job as required for each PR.

Uses [dorny/paths-filter](https://github.com/dorny/paths-filter).

### To test:
1. Ensure that the Doxygen build actually runs on this PR, since we
modified one of the files that was in the filter (the workflow is in the
filter).
2. Ensure that [this
PR](https://github.com/mantidproject/mantid/pull/40507) skips over the
Doxygen build but still passes the doxygen PR check - it's identical to
this one, except I removed the doxygen.yml file from the filter.

*This does not require release notes* because it's a CI tool update.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>