Add arbitrary grouping to BankCalibration (#40511)
In preparation for arbitrary binning, modify `BankCalibration` to mask
detids that are not in the group. `AlignAndFocusPowderSlim` itself was
modified to hard-code the currently used grouping.

Refs #40189 

### To test:

This is a preparatory PR for allowing the grouping information to be
supplied to the system. There are additional unit tests, but the
functionality of `AlignAndFocusPowderSlim` is unchanged.

*This does not require release notes* because it is an internal change.