Move doxygen-publish work into main doxygen workflow (#40519)
Rather than build a second time, combine the rest of the workflow and
only publish the `main` branch. Not publishing `release-next` is
intentional. There is only one site and `main` is the preferred version.

Additional small changes:
- Install pixi environment with --frozen option to avoid updating during
install
- Remove pinning of pixi version to be consistent with other workflows.
Workflows run often enough to identify changes in pixi quickly.
- Update `check-for-changes` to include hand-written pre-compiled
headers as cpp files

This is a follow-on to #40501 and part of #39497.

### To test:

Look at the doxygen workflow and see that the docs are built and the
publish step is skipped.

**This does not require release notes** because it is updating the
workflow to use less compute by reusing steps.